Remove code from `__init__.py`
Empty `__init__.py` are fine, even preferred:
https://docs.python.org/3/tutorial/modules.html#packages

Better to have business logic in   `__init__.py`, only imports
and other "infrastructure" actions.
